Do not exit whole program on error
This gives user a chace to recover from failure. Errors now results in
message being printed to screen as well as proper return code being
passed to caller. This way 'use' function could be used the way that end
of available input might be catched by user script.

Solution must yet be found for situation where user is not interested in
return code, but rather want to close the program immediately (usually
the case).